requirements.txt
httplib2==0.10.3
riotwatcher==2.2.0
aiohttp==2.3.1
websockets==4.0.1
discord.py==0.16.12
